Refactor Chapter 5 Scripts and Update Dataset Preparation and cleaned up unnecessary code and improved documentation for clarity.
