+++ code/chapter05/eval.py
@@ -6,8 +6,8 @@ Provides evaluation on three axes:
     3. **Toy golden set** - Simple Q&A pairs to sanity-check model behavior.
 
 Also includes loss/perplexity computation on held-out JSONL data, and report
-generation (JSON + Markdown). Used by ``scripts/listing_5_4_evaluate.py``
-(Listing 5.4) to compare base model vs. adapter variants.
+generation (JSON + Markdown). Used by ``scripts/listing_5_3_evaluate.py``
+(Listing 5.3) to compare base model vs. adapter variants.
 """
 from __future__ import annotations
 
@@ -473,7 +473,7 @@ def write_report(path: str | Path, obj: Dict[str, Any]) -> None:
     """Write an evaluation results dict as a JSON file.
 
     The JSON report is the machine-readable counterpart to the human-readable
-    Markdown summary generated by ``listing_5_4_evaluate.py``. Both are saved
+    Markdown summary generated by ``listing_5_3_evaluate.py``. Both are saved
     to the same output directory (e.g., ``chapter05/runs/eval_report/``).
 
     Args:
