+++ code/chapter05/examples/README_INTERPRETING_RESULTS.md
@@ -1,6 +1,6 @@
 # Understanding Your Evaluation Results
 
-This guide helps you interpret the evaluation report from `listing_5_4_evaluate.py`.
+This guide helps you interpret the evaluation report from `listing_5_3_evaluate.py`.
 
 ---
