+++ code/chapter05/examples/example_qlora_evaluation_output.md
@@ -5,7 +5,7 @@ This file captures a typical run of the evaluation script when comparing the **b
 ## Command
 
 ```bash
-python chapter05/scripts/listing_5_4_evaluate.py \
+python chapter05/scripts/listing_5_3_evaluate.py \
   --base Qwen/Qwen3-4B-Instruct-2507 \
   --adapter chapter05/runs/dolly_lora \
   --adapter_alt chapter05/runs/dolly_qlora \
