+++ code/chapter05/scripts/fix_safety_complete.sh
@@ -81,7 +81,7 @@ echo "Step 4/4: Evaluating and comparing results..."
 echo "⏱  Time: 5-10 minutes"
 echo ""
 
-python chapter05/scripts/listing_5_4_evaluate.py \
+python chapter05/scripts/listing_5_3_evaluate.py \
   --base Qwen/Qwen3-4B-Instruct-2507 \
   --adapter chapter05/runs/dolly_lora \
   --adapter_alt chapter05/runs/dolly_lora_with_safety \
