+++ code/chapter05/scripts/fix_safety_regression.ps1
@@ -48,7 +48,7 @@ Write-Host "Step 2/2: Evaluating both adapters (r=16 vs r=8)..." -ForegroundColo
 Write-Host "⏱  Estimated time: 5-10 minutes" -ForegroundColor Gray
 Write-Host ""
 
-python chapter05/scripts/listing_5_4_evaluate.py `
+python chapter05/scripts/listing_5_3_evaluate.py `
   --base Qwen/Qwen3-4B-Instruct-2507 `
   --adapter chapter05/runs/dolly_lora `
   --adapter_alt chapter05/runs/dolly_lora_r8 `
